pyproject.toml
[tool.poetry]
name = "api.rhizome.dev"
version = "0.1.0"
description = ""
authors = ["Brian Li <brian@brianli.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.73.0"
rich = "^11.0.0"
iconsdk = "^2.0.0"
python-dotenv = "^0.19.2"
uvicorn = "^0.17.5"
httpx = "^0.22.0"
Pillow = "^9.0.1"
Jinja2 = "^3.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
